+++ CHANGES.rst
@@ -18,11 +18,19 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 - Fix redirections to URLs with host given as IP-literal with brackets.
   Fixes `#1191 <https://github.com/zopefoundation/Zope/issues/1191>`_.
 
-- Fix authentication error viewing ZMI with a user defined outside of zope root.
-  Fixes `#1195 <https://github.com/zopefoundation/Zope/issues/1195>`_.
+- Introduce the decorator ``ZPublisher.zpublish`` to explicitly
+  control publication by ``ZPublisher``.
+  For details see
+  `#1197 <https://github.com/zopefoundation/Zope/pull/1197>`_.
 
 - Fix ``Content-Disposition`` filename for clients without rfc6266 support.
   (`#1198 <https://github.com/zopefoundation/Zope/pull/1198>`_)
+
+- Support ``Chameleon`` ``structure`` expression type.
+  Fixes `#1077 <https://github.com/zopefoundation/Zope/issues/1077>`_.
+
+- Fix authentication error viewing ZMI with a user defined outside of zope root.
+  Fixes `#1195 <https://github.com/zopefoundation/Zope/issues/1195>`_.
 
 
 5.9 (2023-11-24)